done channel refactor

+++ server/src/http/models/myfile.cr
@@ -9,10 +9,9 @@ module DMACServer
       property project : Project
       property size : UInt64
       property modified_at : Time
-      property fileType : String
-      property hidden : Bool
-      property readonly : Bool
-      property group : String
+      property file_type : String
+      property access : Int32
+      property true_access : Int32
 
       raise "No root setup" unless ENV.has_key?("DMAC_ROOT")
       @@root = ENV["DMAC_ROOT"]
@@ -28,6 +27,12 @@ module DMACServer
       @@typeMap[".gif"] = "image"
       @@typeMap[".bmp"] = "image"
       @@typeMap[".pdf"] = "pdf"
+      @@typeMap[".doc"] = "pdf"
+      @@typeMap[".docx"] = "pdf"
+      @@typeMap[".xls"] = "pdf"
+      @@typeMap[".xlsx"] = "pdf"
+      @@typeMap[".ppt"] = "pdf"
+      @@typeMap[".pptx"] = "pdf"
       @@typeMap[".txt"] = "text"
       @@typeMap[".md"] = "text"
       @@typeMap[".r"] = "code"
@@ -56,22 +61,22 @@ module DMACServer
       @@ignore[".git"] = true
       @@ignore[".gitignore"] = true
 
-      def initialize(@project, @data_path)
+      def initialize(@project, @data_path, parent = nil)
         @full_path = @@root + "/" + @project.path.to_s + @data_path
         @rel_path = @project.path.to_s + @data_path
         raise "No such path" unless File.exists?(@full_path)
 
         @type = "folder"
-        @fileType = "folder"
+        @file_type = "folder"
         if File.file?(@full_path)
           @type = "file"
-          @fileType = "unknown"
+          @file_type = "unknown"
           base = File.basename(full_path).downcase
           ext = File.extname(full_path).downcase
           if @@typeMap.has_key?(base)
-            @fileType = @@typeMap[base]
+            @file_type = @@typeMap[base]
           elsif @@typeMap.has_key?(ext)
-            @fileType = @@typeMap[ext]
+            @file_type = @@typeMap[ext]
           end
           @size = File.size(@full_path)
         else
@@ -81,23 +86,9 @@ module DMACServer
         @name = File.basename(@full_path)
         @modified_at = File.stat(@full_path).mtime
 
-        @hidden = false
-        @readonly = false
-        @group = ""
-        return
-
-        chain = @data_path.split("/")
-        chain.each do |s|
-          if s[0] == '.'
-            @hidden = true
-          elsif s[0] == '~'
-            @readonly = true
-          elsif @group == "" && @type == "folder"
-            ss = s.split("_")
-            next if ss.size < 3 || ss[0] != ""
-            @group = ss[1]
-          end
-        end
+        @access = MyFile.get_access(@project, @data_path)
+        @true_access = 0
+        @true_access = get_true_access(parent)
       end
 
       def to_json(read_text = false, public_url = "")
@@ -105,12 +96,13 @@ module DMACServer
           str << "{"
           str << "\"projectId\":\"" << @project.id << "\","
           str << "\"type\":\"" << @type << "\","
-          str << "\"fileType\":\"" << @fileType << "\","
+          str << "\"fileType\":\"" << @file_type << "\","
           str << "\"name\":\"" << @name << "\","
           str << "\"dataPath\":\"" << @data_path << "\","
           str << "\"publicUrl\":\"" << public_url << "\","
           str << "\"size\":\"" << @size << "\","
-          str << "\"readonly\":" << @readonly << ","
+          str << "\"access\":" << @access << ","
+          str << "\"trueAccess\":" << @true_access << ","
           if read_text
             str << "\"text\":" << get_text.to_json << ","
           else
@@ -130,21 +122,52 @@ module DMACServer
         File.write(@full_path, text)
       end
 
+      def get_true_access(parent)
+        return 2 if @access == 2
+        access = 0
+        if parent.nil?
+          dp = ""
+          @data_path.split("/") do |s|
+            dp += "/" + s
+            a = MyFile.get_access(@project, dp)
+            access = Math.max(access, a)
+            break if access == 2
+          end
+        else
+          access = parent.true_access
+        end
+        return Math.max(access, @access)
+      end
+
       def viewable?(control)
         role = control.role.to_s
-        group = control.group_name.to_s
         return true if role == "Owner" || role == "Admin"
-        return false if @hidden
-        return true if group == "" || @group == ""
-        return group == @group
+        return @true_access != 2
       end
 
       def editable?(control)
         role = control.role.to_s
         return true if role == "Owner" || role == "Admin"
         return false if role == "Viewer"
-        return false if !viewable?(control)
-        return !@readonly
+        return false if @file_type == "folder"
+        return @true_access == 0
+      end
+
+      def can_add_file?(control, filename)
+        role = control.role.to_s
+        return false unless @file_type == "folder"
+        return true if role == "Owner" || role == "Admin"
+        return false if role == "Viewer"
+        return false if @true_access > 0
+        target_path = @full_path + "/" + filename
+        return true unless File.exists?(target_path)
+        target_data_path = @data_path + "/" + filename
+        target_file = MyFile.new(@project, target_data_path, self)
+        return target_file.editable?(control)
+      end
+
+      def self.get_access(project, data_path)
+        return 0
       end
 
       def self.collect_files(control, project, data_path)
@@ -157,10 +180,17 @@ module DMACServer
           next if @@ignore.has_key? filename.to_s
           dp = "/" + filename
           dp = data_path + dp unless data_path == "/"
-          file = MyFile.new(project, dp)
+          file = MyFile.new(project, dp, dir)
           files << file if file.viewable?(control)
         end
         return files
+      end
+
+      def self.get_parent_data_path(data_path)
+        index = data_path.rindex("/")
+        parent_data_path = "/"
+        parent_data_path = data_path[0, index] unless index.nil? || data_path == "/"
+        return parent_data_path
       end
 
       def self.create_folder(project, data_path)
@@ -168,23 +198,25 @@ module DMACServer
         Dir.mkdir(full_path)
       end
 
+      def self.create_project_folder!(user, project_name)
+        full_path = @@root + "/" + user.username.to_s + "/" + project_name
+        Dir.mkdir(full_path)
+      end
+
       def self.create_file(project, data_path, control)
-        raise "No permission" unless MyFile.check_parent(project, data_path, control)
         full_path = @@root + "/" + project.path.to_s + data_path
         raise "Already exist" if File.exists?(full_path)
+        role = control.role.to_s
+        if role == "Editor"
+          parent_data_path = MyFile.get_parent_data_path(data_path)
+          parent_file = MyFile.new(project, parent_data_path)
+          raise "File permission denied" if parent_file.true_access > 0
+        end
         File.write(full_path, "")
       end
 
-      def self.check_parent(project, data_path, control)
-        index = data_path.rindex("/")
-        parent_data_path = "/"
-        parent_data_path = data_path[0, index] unless index.nil? || data_path == "/"
-        file = MyFile.new(project, parent_data_path)
-        return file.editable?(control)
-      end
-
-      def self.delete_folder(project, data_path)
-        full_path = @@root + "/" + project.path.to_s + data_path
+      def self.delete_project_folder(project)
+        full_path = @@root + "/" + project.path.to_s
         MyFile.delete_files(full_path)
       end
 
@@ -219,10 +251,6 @@ module DMACServer
       end
 
       def self.upload_file(project, data_path, file, control)
-        dir = MyFile.new(project, data_path)
-        raise "Target is not a folder" unless dir.type == "folder"
-        raise "No permission" unless dir.editable?(control)
-
         full_path = @@root + "/" + project.path.to_s
         prefix_length = full_path.size + 1
         full_path = full_path + data_path
@@ -230,61 +258,15 @@ module DMACServer
         filename = file.filename
         raise "No filename included in upload" if !filename.is_a?(String)
 
-        target_path = MyFile.make_target(full_path, filename, true)
+        dir = MyFile.new(project, data_path)
+        raise "No file permission" unless dir.can_add_file?(control, filename)
+
+        target_path = full_path + "/" + filename
         File.open(target_path, "w") do |f|
           IO.copy(file.tmpfile, f)
         end
 
         return target_path[prefix_length..-1]
-      end
-
-      def self.make_target(path, name, upload)
-        new_name = String.build do |str|
-          name.chars.each do |c|
-            ord = c.ord
-            if ord == 46
-              str << c
-            elsif ord >= 48 && ord <= 57
-              str << c
-            elsif ord >= 65 && ord <= 90
-              str << c
-            elsif ord == 95
-              str << c
-            elsif ord >= 97 && ord <= 122
-              str << c
-            elsif ord == 126
-              str << c
-            else
-              str << "_"
-            end
-          end
-        end
-        raise "cannot make file name" if new_name == ""
-
-        if (upload && ((new_name.starts_with? '.') || (new_name.starts_with? '~')))
-          new_name = "_" + new_name.lchop
-        end
-        if (new_name.starts_with? '-')
-          new_name = "_" + new_name.lchop
-        end
-        if ((new_name.ends_with? '.') || (new_name.ends_with? '-'))
-          new_name = new_name.rchop + "_"
-        end
-
-        filenames = {} of String => Bool
-        Dir.foreach path do |filename|
-          filenames[filename.to_s] = true
-        end
-
-        i = 0
-        ext = File.extname(new_name)
-        base = File.basename(new_name, ext)
-        while filenames.has_key?(new_name)
-          i = i + 1
-          new_name = base + "_" + i.to_s + ext
-        end
-
-        return path + "/" + new_name
       end
 
       def self.get_download_path(project, download)
@@ -300,27 +282,25 @@ module DMACServer
       def self.save_text_file(project, data_path, text, control)
         file = MyFile.new(project, data_path)
         raise "No permission" unless file.editable?(control)
-        raise "Not text file" unless file.fileType == "text" || file.fileType == "code"
+        raise "Not text file" unless file.file_type == "text" || file.file_type == "code"
         file.save_text(text)
       end
 
-      def self.copy_files(source_files, target_file, source_control, target_control)
-        raise "Target is not a folder" unless target_file.type == "folder"
-        raise "No permission" unless target_file.editable?(target_control)
+      def self.copy_files(source_files, target_path, target_control)
         new_folders = {} of String => Bool
         source_files.each do |f|
-          source_project = f.project
           if f.type == "file"
-            MyFile.copy_file(f.full_path, target_file.full_path, source_control, source_project)
+            MyFile.copy_file(f.full_path, target_path, target_control)
           else
-            MyFile.copy_folder(f.full_path, target_file.full_path, new_folders, source_control, source_project)
+            MyFile.copy_folder(f.full_path, target_path, target_control, new_folders)
           end
         end
       end
 
-      def self.copy_file(source_full_path, target_full_path, control, project)
-        return unless MyFile.check_source(control, project, source_full_path)
-        new_full_path = MyFile.make_name(source_full_path, target_full_path)
+      def self.copy_file(source_full_path, target_path, target_control)
+        return unless MyFile.check_copy(source_full_path, target_path, target_control)
+        filename = File.basename(source_full_path)
+        new_full_path = target_path.full_path + "/" + filename
         File.open(new_full_path, "w") do |tf|
           File.open(source_full_path) do |sf|
             IO.copy(sf, tf)
@@ -328,46 +308,36 @@ module DMACServer
         end
       end
 
-      def self.copy_folder(source_full_path, target_full_path, new_folders, control, project)
-        return unless MyFile.check_source(control, project, source_full_path)
-        new_target_full_path = MyFile.make_name(source_full_path, target_full_path)
-        Dir.mkdir(new_target_full_path)
-        new_folders[new_target_full_path] = true
+      def self.copy_folder(source_full_path, target_path, target_control, new_folders)
+        return unless MyFile.check_copy(source_full_path, target_path, target_control)
+        folder_name = File.basename(source_full_path)
+        new_target_full_path = target_path.full_path + "/" + folder_name
+        Dir.mkdir(new_target_full_path) unless File.exists? new_target_full_path
+        new_target_data_path = target_path.data_path + "/" + folder_name
+        new_target_path = MyFile.new(target_path.project, new_target_data_path, target_path)
+        new_folders[new_target_path.full_path] = true
+
         Dir.foreach source_full_path do |filename|
           next if @@ignore.has_key? filename.to_s
           new_source_full_path = source_full_path + "/" + filename
           if File.file? new_source_full_path
-            MyFile.copy_file(new_source_full_path, new_target_full_path, control, project)
+            MyFile.copy_file(new_source_full_path, new_target_path, target_control)
           elsif (File.directory? new_source_full_path) && (!new_folders.has_key?(new_source_full_path))
-            MyFile.copy_folder(new_source_full_path, new_target_full_path, new_folders, control, project)
+            MyFile.copy_folder(new_source_full_path, new_target_path, target_control, new_folders)
           end
         end
       end
 
-      def self.check_source(control, project, full_path)
-        project_root = @@root + "/" + project.path.to_s
-        rel_path = full_path[project_root.size..-1]
-        data_path = rel_path
-        file = MyFile.new(project, data_path)
-        return file.viewable?(control)
-      end
+      def self.check_copy(source_full_path, target_path, target_control)
+        return false unless target_control.role.to_s != "Editor" || File.file? source_full_path
 
-      def self.make_name(source_name, target_path)
-        filenames = {} of String => Bool
-        Dir.foreach target_path do |filename|
-          filenames[filename.to_s] = true
-        end
+        filename = File.basename(source_full_path)
+        new_full_path = target_path.full_path + "/" + filename
+        return true unless File.exists? new_full_path
 
-        i = 0
-        new_name = File.basename(source_name)
-        ext = File.extname(source_name)
-        base = File.basename(source_name, ext)
-        while filenames.has_key?(new_name)
-          i = i + 1
-          new_name = base + "_" + i.to_s + ext
-        end
-
-        return target_path + "/" + new_name
+        new_data_path = target_path.data_path + "/" + filename
+        new_file = MyFile.new(target_path.project, new_data_path, target_path)
+        return new_file.editable? target_control
       end
 
       def self.unzip_file(project, data_path)
@@ -393,7 +363,8 @@ module DMACServer
       def self.unzip_to_temp(source)
         temp_path = @@tmp + "/" + SecureRandom.hex(32).to_s
         Dir.mkdir(temp_path)
-        command = "unzip " + source + " -d " + temp_path
+        command = "unzip \"" + source + "\" -d " + temp_path
+        puts command
         io = IO::Memory.new
         Process.run(command, shell: true, output: io)
         return temp_path
@@ -401,7 +372,8 @@ module DMACServer
 
       def self.copy_file_from_outside(source, target)
         name = File.basename(source)
-        target_path = MyFile.make_target(target, name, false)
+        target_path = target + "/" + name
+        return if ((File.exists? target_path) && (File.directory? target_path))
         File.open(target_path, "w") do |tf|
           File.open(source) do |sf|
             IO.copy(sf, tf)
@@ -411,8 +383,9 @@ module DMACServer
 
       def self.copy_folder_from_outside(source, target)
         name = File.basename(source)
-        new_target = MyFile.make_target(target, name, false)
-        Dir.mkdir(new_target)
+        new_target = target + "/" + name
+        return if ((File.exists? new_target) && (File.file? new_target))
+        Dir.mkdir(new_target) unless File.exists? new_target
         Dir.foreach source do |filename|
           next if @@ignore.has_key? filename.to_s
           new_source = source + "/" + filename
@@ -426,15 +399,45 @@ module DMACServer
 
       def self.prepare_download(download, project, data_path, control)
         file = MyFile.new(project, data_path)
+        raise "No permission" unless file.viewable?(control)
         return if file.type == "file"
+
         key = download.key.to_s
         temp_path = @@tmp + "/" + key
         Dir.mkdir(temp_path)
         new_folders = {} of String => Bool
-        MyFile.copy_folder(file.full_path, temp_path, new_folders, control, project)
+        MyFile.copy_folder_to_outside(file, control, temp_path, new_folders)
         command = "cd " + temp_path + " && zip -r " + key + " *"
         io = IO::Memory.new
         Process.run(command, shell: true, output: io)
+      end
+
+      def self.copy_folder_to_outside(file, control, path, new_folders)
+        raise "No permission" unless file.viewable?(control)
+        new_full_path = path + "/" + file.name
+        Dir.mkdir(new_full_path) unless File.exists? new_full_path
+        new_folders[new_full_path] = true
+
+        Dir.foreach file.full_path do |filename|
+          next if @@ignore.has_key? filename.to_s
+          new_data_path = file.data_path + "/" + filename
+          new_file = MyFile.new(file.project, new_data_path, file)
+          if new_file.type == "file"
+            MyFile.copy_file_to_outside(new_file, control, new_full_path)
+          elsif !new_folders.has_key?(new_file.full_path)
+            MyFile.copy_folder_to_outside(new_file, control, new_full_path, new_folders)
+          end
+        end
+      end
+
+      def self.copy_file_to_outside(file, control, path)
+        raise "No permission" unless file.viewable?(control)
+        new_full_path = path + "/" + file.name
+        File.open(new_full_path, "w") do |tf|
+          File.open(file.full_path) do |sf|
+            IO.copy(sf, tf)
+          end
+        end
       end
 
       def self.copy_project_files(source, target)
